pyproject.toml
[project]
name = "skillmatch-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "psycopg2>=2.9.10",
]
